+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Note: `search` increased `description` weight from `0.9` to `1.0`.
 
 ## `20241107t132800-all`
  * `search` uses the `IndexedScore` to reduce memory allocations.